+++ packages/uni-app-react/example/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @js-css/uni-app-react-demo
 
+## 0.0.6
+
+### Patch Changes
+
+- feat:
+  1. add event & element create hook
+  2. adj lifecycle to outsite
+
+- Updated dependencies
+  - @js-css/uni-app-react@0.0.6
+
 ## 0.0.5
 
 ### Patch Changes